chore: move Jest tests under `test` dir
